rename base query with retry
