+++ src/store/services/categoryApi.ts
@@ -1,10 +1,10 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
 import { Category } from '../models/interfaces';
-import customBaseQuery from './customBaseQuery';
+import baseQueryWithRetry from './baseQueryWithRetry';
 
 const categoryApi = createApi({
   reducerPath: 'categoryApi',
-  baseQuery: customBaseQuery,
+  baseQuery: baseQueryWithRetry,
   endpoints: (builder) => ({
     fetchCategories: builder.query<Category[], void>({
       query: () => {
